requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate>=4.0
alembic>=1.13
python-nmap==0.7.1
requests==2.31.0
APScheduler==3.10.4
psutil==5.9.6
netifaces==0.11.0
speedtest-cli==2.1.3
pythonping==1.1.4
Pillow==10.0.1